Tests/LibWeb/Text/expected/document-scrollingElement.txt

document.compatMode: CSS1Compat
document.scrollingElement is document root element: true
